Fix for issue #902
In build tracks function, I added a check to see if the browser will
attempt to do present captions natively. (Currently only a feature in
Chrome. Outside of mejs, my if statement would include a check to see
if any textTracks were declared (t.domNode.textTracks.length > 0), but
the buildtracks function already checks for player.tracks.length.) If
this is true, we prefer to use mejs captions and loop through the
textTracks to set their mode to "hidden" to prevent display.